Update variables handled by TTreeReader
There have been some updates to TopAnalysis,
so we add a few more nominal systematics.

Also move some variables that are not vanilla
TopAnalysis output (that we add, e.g. the
truthMatched lepton variable) out of AnaBase
so that they can be part of AIDA-Loops LoopAlg
(this is to stay consistent with the original
design of TopLoop).