Use true_isPrompt from v23+ ntuples, cmake flag for older
This commit makes use of the true_isPrompt the default variable for
checking MC truth information about truth leptons

A flag has been added to tell cmake to build with backwards
compatibility: `-DSgTopNtuple_v22=ON` will turn on the compiler
definiton `SGTOPNTUPLE_V22`. This is a public TopLoop definition so it
propagates to dependent algorithms.