Use ntuple version determination for campaign weighting
Unfortunately, with different ntuple versions we have different total
lumis and therefore campaign weighting is dependent on the single top
ntuple version used. This commit makes the meta service class store an
ntuple version, it defaults to v23 but if the getNtupleVersion
function is called the internal version is updated. This ntuple
version stored in the class is used when retrieving campaign weights

It seems a bit complicated to handle two ntuple versions, but of
course we'll be adding more in the future and this code should make
that less of a headache!